Created a factories folder for factory code
