+++ include/factories/UterineRegularCellFactory.hpp
@@ -1,23 +1,23 @@
-#ifndef INCLUDE_UTERINESIMPLECELLFACTORY_HPP_
-#define INCLUDE_UTERINESIMPLECELLFACTORY_HPP_
+#ifndef INCLUDE_UTERINEREGULARCELLFACTORY_HPP_
+#define INCLUDE_UTERINEREGULARCELLFACTORY_HPP_
 
 #include <iostream>
 #include <string>
-#include "../include/toml.hpp"
+#include "../toml.hpp"
 #include "AbstractUterineCellFactory.hpp"
 #include "MonodomainProblem.hpp"
-#include "SimpleStimulus.hpp"
+#include "RegularStimulus.hpp"
 
-class UterineSimpleCellFactory : public AbstractUterineCellFactory {
+class UterineRegularCellFactory : public AbstractUterineCellFactory {
  private:
   double mpX_stim_start;
   double mpX_stim_end;
   double mpY_stim_start;
   double mpY_stim_end;
-  boost::shared_ptr<SimpleStimulus> mpStimulus;
+  boost::shared_ptr<RegularStimulus> mpStimulus;
 
  public:
-  UterineSimpleCellFactory();
+  UterineRegularCellFactory();
   AbstractCvodeCell* CreateCardiacCellForTissueNode(Node<2>* pNode);
   void ReadParams(std::string general_param_file);
   void ReadCellParams(std::string cell_param_file);
@@ -25,4 +25,5 @@ class UterineSimpleCellFactory : public AbstractUterineCellFactory {
   void WriteLogInfo(std::string log_file);
 };
 
-#endif  // INCLUDE_UTERINESIMPLECELLFACTORY_HPP_
+#endif  // INCLUDE_UTERINEREGULARCELLFACTORY_HPP_
+
