+++ src/factories/AbstractUterineCellFactory3d.cpp
@@ -1,14 +1,14 @@
-#include "../include/AbstractUterineCellFactory.hpp"
+#include "../../include/factories/AbstractUterineCellFactory3d.hpp"
 #include "Exception.hpp"
 
-AbstractUterineCellFactory::AbstractUterineCellFactory() :
-  AbstractCardiacCellFactory<2>() {
-  ReadParams(USMC_2D_SYSTEM_CONSTANTS::GENERAL_PARAM_FILE);
+AbstractUterineCellFactory3d::AbstractUterineCellFactory3d() : 
+  AbstractCardiacCellFactory<3>() {
+  ReadParams(USMC_3D_SYSTEM_CONSTANTS::GENERAL_PARAM_FILE);
 }
 
 
-AbstractCvodeCell* AbstractUterineCellFactory::CreateCardiacCellForTissueNode(
-  Node<2>* pNode) {
+AbstractCvodeCell* AbstractUterineCellFactory3d::CreateCardiacCellForTissueNode(
+  Node<3>* pNode) {
   AbstractCvodeCell* cell;
 
   switch (mpCell_id) {
@@ -25,8 +25,8 @@ AbstractCvodeCell* AbstractUterineCellFactory::CreateCardiacCellForTissueNode(
       cell = new CellMeans2023FromCellMLCvode(mpSolver, mpZeroStimulus);
 
       for (auto it=mpCell_parameters.begin();
-        it != mpCell_parameters.end();
-        ++it) {
+          it != mpCell_parameters.end();
+          ++it) {
         cell->SetParameter(it->first, it->second);
       }
       break;
@@ -35,8 +35,8 @@ AbstractCvodeCell* AbstractUterineCellFactory::CreateCardiacCellForTissueNode(
       cell = new CellTong2014FromCellMLCvode(mpSolver, mpZeroStimulus);
 
       for (auto it=mpCell_parameters.begin();
-        it != mpCell_parameters.end();
-        ++it) {
+          it != mpCell_parameters.end();
+          ++it) {
         cell->SetParameter(it->first, it->second);
       }
       break;
@@ -45,8 +45,8 @@ AbstractCvodeCell* AbstractUterineCellFactory::CreateCardiacCellForTissueNode(
       cell = new CellRoesler2024FromCellMLCvode(mpSolver, mpZeroStimulus);
 
       for (auto it=mpCell_parameters.begin();
-        it != mpCell_parameters.end();
-        ++it) {
+          it != mpCell_parameters.end();
+          ++it) {
         cell->SetParameter(it->first, it->second);
       }
       break;
@@ -60,12 +60,12 @@ AbstractCvodeCell* AbstractUterineCellFactory::CreateCardiacCellForTissueNode(
 }
 
 
-std::string AbstractUterineCellFactory::GetCellType() {
+std::string AbstractUterineCellFactory3d::GetCellType() {
   return mpCell_type;
 }
 
 
-std::string AbstractUterineCellFactory::GetCellParamFile() {
+std::string AbstractUterineCellFactory3d::GetCellParamFile() {
   std::string cell_param_file = "";
 
   if (mpEstrus == "") {
@@ -78,32 +78,29 @@ std::string AbstractUterineCellFactory::GetCellParamFile() {
   return cell_param_file;
 }
 
-
-void AbstractUterineCellFactory::ReadParams(std::string general_param_file) {
-  std::string general_param_path = USMC_2D_SYSTEM_CONSTANTS::CONFIG_DIR +
+void AbstractUterineCellFactory3d::ReadParams(std::string general_param_file) {
+  std::string general_param_path = USMC_3D_SYSTEM_CONSTANTS::CONFIG_DIR +
     general_param_file;
   const auto params = toml::parse(general_param_path);
 
   mpCell_type = toml::find<std::string>(params, "cell_type");
 
-  // Get the cell configuration
-  std::string cell_param_file = "";
-
   if (mpCell_type == std::string("Roesler")) {
     // Get the estrus phase as well
-    const std::string estrus_phase = toml::find<std::string>(params,
-      "estrus");
-    cell_param_file = mpCell_type + "_" + estrus_phase + ".toml";
+    mpEstrus = toml::find<std::string>(params, "estrus");
   } else {
-    cell_param_file = mpCell_type + ".toml";
+    mpEstrus = "";
   }
+
+  // Get the cell configuration
+  std::string cell_param_file = GetCellParamFile();
 
   ReadCellParams(cell_param_file);
 }
 
 
-void AbstractUterineCellFactory::ReadCellParams(std::string cell_param_file) {
-  std::string cell_param_path = USMC_2D_SYSTEM_CONSTANTS::CONFIG_DIR +
+void AbstractUterineCellFactory3d::ReadCellParams(std::string cell_param_file) {
+  std::string cell_param_path = USMC_3D_SYSTEM_CONSTANTS::CONFIG_DIR +
     cell_param_file;
   const auto cell_params = toml::parse(cell_param_path);
 
@@ -117,9 +114,9 @@ void AbstractUterineCellFactory::ReadCellParams(std::string cell_param_file) {
 }
 
 
-void AbstractUterineCellFactory::PrintParams() {
-  std::cout << "mpCell_type = " << mpCell_type << "\n";
-  std::cout << "mpCell_id = " << mpCell_id << "\n";
+void AbstractUterineCellFactory3d::PrintParams() {
+  std::cout << "mpCell_type = " << mpCell_type << std::endl;
+  std::cout << "mpCell_id = " << mpCell_id << std::endl;
   std::cout << "mpCell_parameters\n";
 
   for (auto it=mpCell_parameters.begin(); it != mpCell_parameters.end(); ++it) {
@@ -128,7 +125,7 @@ void AbstractUterineCellFactory::PrintParams() {
 }
 
 
-void AbstractUterineCellFactory::WriteLogInfo(std::string log_file) {
+void AbstractUterineCellFactory3d::WriteLogInfo(std::string log_file) {
   std::ofstream log_stream;
   log_stream.open(log_file, ios::app);  // Open log file in append mode
 
@@ -140,3 +137,4 @@ void AbstractUterineCellFactory::WriteLogInfo(std::string log_file) {
 
   log_stream.close();
 }
+
