+++ src/factories/UterineRegularCellFactory.cpp
@@ -1,25 +1,23 @@
-#include "../include/UterineRegularCellFactory3d.hpp"
+#include "../../include/factories/UterineRegularCellFactory.hpp"
 #include "Exception.hpp"
 
-UterineRegularCellFactory3d::UterineRegularCellFactory3d() :
-  AbstractUterineCellFactory3d(),
+UterineRegularCellFactory::UterineRegularCellFactory() :
+  AbstractUterineCellFactory(),
   mpStimulus(new RegularStimulus(0.0, 0.0, 0.1, 0.0)) {
-  ReadParams(USMC_3D_SYSTEM_CONSTANTS::GENERAL_PARAM_FILE);
-  ReadCellParams(AbstractUterineCellFactory3d::GetCellParamFile());
+  ReadParams(USMC_2D_SYSTEM_CONSTANTS::GENERAL_PARAM_FILE);
+  ReadCellParams(AbstractUterineCellFactory::GetCellParamFile());
 }
 
 
-AbstractCvodeCell* UterineRegularCellFactory3d::CreateCardiacCellForTissueNode(
-  Node<3>* pNode) {
+AbstractCvodeCell* UterineRegularCellFactory::CreateCardiacCellForTissueNode(
+  Node<2>* pNode) {
   double x = pNode->rGetLocation()[0];
   double y = pNode->rGetLocation()[1];
-  double z = pNode->rGetLocation()[2];
 
   AbstractCvodeCell* cell;
 
   if (x >= mpX_stim_start && x <= mpX_stim_end &&
-      y >= mpY_stim_start && y <= mpY_stim_end &&
-      z >= mpZ_stim_start && z <= mpZ_stim_end) {
+      y >= mpY_stim_start && y <= mpY_stim_end) {
     switch (mpCell_id) {
       case 0:
         cell = new CellHodgkinHuxley1952FromCellMLCvode(mpSolver,
@@ -68,15 +66,15 @@ AbstractCvodeCell* UterineRegularCellFactory3d::CreateCardiacCellForTissueNode(
     return cell;
   } else {
     /* The other cells have zero stimuli. */
-    return AbstractUterineCellFactory3d::CreateCardiacCellForTissueNode(pNode);
+    return AbstractUterineCellFactory::CreateCardiacCellForTissueNode(pNode);
   }
 }
 
 
-void UterineRegularCellFactory3d::ReadParams(std::string general_param_file) {
-  AbstractUterineCellFactory3d::ReadParams(general_param_file);
+void UterineRegularCellFactory::ReadParams(std::string general_param_file) {
+  AbstractUterineCellFactory::ReadParams(general_param_file);
 
-  std::string general_param_path = USMC_3D_SYSTEM_CONSTANTS::CONFIG_DIR +
+  std::string general_param_path = USMC_2D_SYSTEM_CONSTANTS::CONFIG_DIR +
     general_param_file;
   const auto params = toml::parse(general_param_path);
 
@@ -85,12 +83,11 @@ void UterineRegularCellFactory3d::ReadParams(std::string general_param_file) {
   mpX_stim_end = toml::find<double>(params, "x_stim_end");
   mpY_stim_start = toml::find<double>(params, "y_stim_start");
   mpY_stim_end = toml::find<double>(params, "y_stim_end");
-  mpZ_stim_start = toml::find<double>(params, "z_stim_start");
-  mpZ_stim_end = toml::find<double>(params, "z_stim_end");
 }
 
-void UterineRegularCellFactory3d::ReadCellParams(std::string cell_param_file) {
-  std::string cell_param_path = USMC_3D_SYSTEM_CONSTANTS::CONFIG_DIR +
+
+void UterineRegularCellFactory::ReadCellParams(std::string cell_param_file) {
+  std::string cell_param_path = USMC_2D_SYSTEM_CONSTANTS::CONFIG_DIR +
     cell_param_file;
   const auto cell_params = toml::parse(cell_param_path);
 
@@ -102,14 +99,12 @@ void UterineRegularCellFactory3d::ReadCellParams(std::string cell_param_file) {
 }
 
 
-void UterineRegularCellFactory3d::PrintParams() {
-  AbstractUterineCellFactory3d::PrintParams();
+void UterineRegularCellFactory::PrintParams() {
+  AbstractUterineCellFactory::PrintParams();
   std::cout << "mpX_stim_start = " << mpX_stim_start << "\n";
   std::cout << "mpX_stim_end = " << mpX_stim_end << "\n";
   std::cout << "mpY_stim_start = " << mpY_stim_start << "\n";
   std::cout << "mpY_stim_end = " << mpY_stim_end << "\n";
-  std::cout << "mpZ_stim_start = " << mpZ_stim_start << "\n";
-  std::cout << "mpZ_stim_end = " << mpZ_stim_end << "\n";
   std::cout << "mpStimulus magnitude = "
     << mpStimulus->GetMagnitude()
     << std::endl;
@@ -121,8 +116,8 @@ void UterineRegularCellFactory3d::PrintParams() {
 }
 
 
-void UterineRegularCellFactory3d::WriteLogInfo(std::string log_file) {
-  AbstractUterineCellFactory3d::WriteLogInfo(log_file);
+void UterineRegularCellFactory::WriteLogInfo(std::string log_file) {
+  AbstractUterineCellFactory::WriteLogInfo(log_file);
 
   std::ofstream log_stream;
   log_stream.open(log_file, ios::app);  // Open log file in append mode
@@ -144,9 +139,7 @@ void UterineRegularCellFactory3d::WriteLogInfo(std::string log_file) {
   log_stream << "  period: " << mpStimulus->GetPeriod() << " ms" << std::endl;
   log_stream << "  stimulated region: " << mpX_stim_start << " <= x <= ";
   log_stream << mpX_stim_end << "   " << mpY_stim_start << " <= y <= ";
-  log_stream << mpY_stim_end << "   " << mpZ_stim_start << " <= z <= ";
-  log_stream << mpZ_stim_end << std::endl;
+  log_stream << mpY_stim_end << std::endl;
 
   log_stream.close();
 }
-
