+++ src/factories/UterineRegularCellFactory3d.cpp
@@ -1,23 +1,25 @@
-#include "../include/UterineSimpleCellFactory.hpp"
+#include "../../include/factories/UterineRegularCellFactory3d.hpp"
 #include "Exception.hpp"
 
-UterineSimpleCellFactory::UterineSimpleCellFactory() :
-  AbstractUterineCellFactory(),
-  mpStimulus(new SimpleStimulus(0.0, 0.0)) {
-  ReadParams(USMC_2D_SYSTEM_CONSTANTS::GENERAL_PARAM_FILE);
-  ReadCellParams(AbstractUterineCellFactory::GetCellParamFile());
+UterineRegularCellFactory3d::UterineRegularCellFactory3d() :
+  AbstractUterineCellFactory3d(),
+  mpStimulus(new RegularStimulus(0.0, 0.0, 0.1, 0.0)) {
+  ReadParams(USMC_3D_SYSTEM_CONSTANTS::GENERAL_PARAM_FILE);
+  ReadCellParams(AbstractUterineCellFactory3d::GetCellParamFile());
 }
 
 
-AbstractCvodeCell* UterineSimpleCellFactory::CreateCardiacCellForTissueNode(
-  Node<2>* pNode) {
+AbstractCvodeCell* UterineRegularCellFactory3d::CreateCardiacCellForTissueNode(
+  Node<3>* pNode) {
   double x = pNode->rGetLocation()[0];
   double y = pNode->rGetLocation()[1];
+  double z = pNode->rGetLocation()[2];
 
   AbstractCvodeCell* cell;
 
   if (x >= mpX_stim_start && x <= mpX_stim_end &&
-      y >= mpY_stim_start && y <= mpY_stim_end) {
+      y >= mpY_stim_start && y <= mpY_stim_end &&
+      z >= mpZ_stim_start && z <= mpZ_stim_end) {
     switch (mpCell_id) {
       case 0:
         cell = new CellHodgkinHuxley1952FromCellMLCvode(mpSolver,
@@ -66,15 +68,15 @@ AbstractCvodeCell* UterineSimpleCellFactory::CreateCardiacCellForTissueNode(
     return cell;
   } else {
     /* The other cells have zero stimuli. */
-    return AbstractUterineCellFactory::CreateCardiacCellForTissueNode(pNode);
+    return AbstractUterineCellFactory3d::CreateCardiacCellForTissueNode(pNode);
   }
 }
 
 
-void UterineSimpleCellFactory::ReadParams(std::string general_param_file) {
-  AbstractUterineCellFactory::ReadParams(general_param_file);
+void UterineRegularCellFactory3d::ReadParams(std::string general_param_file) {
+  AbstractUterineCellFactory3d::ReadParams(general_param_file);
 
-  std::string general_param_path = USMC_2D_SYSTEM_CONSTANTS::CONFIG_DIR +
+  std::string general_param_path = USMC_3D_SYSTEM_CONSTANTS::CONFIG_DIR +
     general_param_file;
   const auto params = toml::parse(general_param_path);
 
@@ -83,30 +85,35 @@ void UterineSimpleCellFactory::ReadParams(std::string general_param_file) {
   mpX_stim_end = toml::find<double>(params, "x_stim_end");
   mpY_stim_start = toml::find<double>(params, "y_stim_start");
   mpY_stim_end = toml::find<double>(params, "y_stim_end");
+  mpZ_stim_start = toml::find<double>(params, "z_stim_start");
+  mpZ_stim_end = toml::find<double>(params, "z_stim_end");
 }
 
-
-void UterineSimpleCellFactory::ReadCellParams(std::string cell_param_file) {
-  std::string cell_param_path = USMC_2D_SYSTEM_CONSTANTS::CONFIG_DIR +
+void UterineRegularCellFactory3d::ReadCellParams(std::string cell_param_file) {
+  std::string cell_param_path = USMC_3D_SYSTEM_CONSTANTS::CONFIG_DIR +
     cell_param_file;
   const auto cell_params = toml::parse(cell_param_path);
 
   // Stimulus parameters
   mpStimulus->SetMagnitude(toml::find<double>(cell_params, "magnitude"));
+  mpStimulus->SetPeriod(toml::find<double>(cell_params, "period"));
   mpStimulus->SetDuration(toml::find<double>(cell_params, "duration"));
   mpStimulus->SetStartTime(toml::find<double>(cell_params, "start_time"));
 }
 
 
-void UterineSimpleCellFactory::PrintParams() {
-  AbstractUterineCellFactory::PrintParams();
+void UterineRegularCellFactory3d::PrintParams() {
+  AbstractUterineCellFactory3d::PrintParams();
   std::cout << "mpX_stim_start = " << mpX_stim_start << "\n";
   std::cout << "mpX_stim_end = " << mpX_stim_end << "\n";
   std::cout << "mpY_stim_start = " << mpY_stim_start << "\n";
   std::cout << "mpY_stim_end = " << mpY_stim_end << "\n";
+  std::cout << "mpZ_stim_start = " << mpZ_stim_start << "\n";
+  std::cout << "mpZ_stim_end = " << mpZ_stim_end << "\n";
   std::cout << "mpStimulus magnitude = "
     << mpStimulus->GetMagnitude()
     << std::endl;
+  std::cout << "mpStimulus period = " << mpStimulus->GetPeriod() << std::endl;
   std::cout << "mpStimulus duration = "
     << mpStimulus->GetDuration()
     << std::endl;
@@ -114,14 +121,14 @@ void UterineSimpleCellFactory::PrintParams() {
 }
 
 
-void UterineSimpleCellFactory::WriteLogInfo(std::string log_file) {
-  AbstractUterineCellFactory::WriteLogInfo(log_file);
+void UterineRegularCellFactory3d::WriteLogInfo(std::string log_file) {
+  AbstractUterineCellFactory3d::WriteLogInfo(log_file);
 
   std::ofstream log_stream;
   log_stream.open(log_file, ios::app);  // Open log file in append mode
 
   log_stream << "Stimulus parameters" << std::endl;
-  log_stream << "  type: simple" << std::endl;
+  log_stream << "  type: regular" << std::endl;
   log_stream << "  start time: "
     << mpStimulus->GetStartTime()
     << " ms"
@@ -134,9 +141,12 @@ void UterineSimpleCellFactory::WriteLogInfo(std::string log_file) {
     << mpStimulus->GetMagnitude()
     << " uA/cm2"
     << std::endl;
+  log_stream << "  period: " << mpStimulus->GetPeriod() << " ms" << std::endl;
   log_stream << "  stimulated region: " << mpX_stim_start << " <= x <= ";
   log_stream << mpX_stim_end << "   " << mpY_stim_start << " <= y <= ";
-  log_stream << mpY_stim_end << std::endl;
+  log_stream << mpY_stim_end << "   " << mpZ_stim_start << " <= z <= ";
+  log_stream << mpZ_stim_end << std::endl;
 
   log_stream.close();
 }
+
