+++ src/factories/UterineSimpleCellFactory3d.cpp
@@ -1,4 +1,4 @@
-#include "../include/UterineSimpleCellFactory3d.hpp"
+#include "../../include/factories/UterineSimpleCellFactory3d.hpp"
 #include "Exception.hpp"
 
 UterineSimpleCellFactory3d::UterineSimpleCellFactory3d() :
