+++ src/factories/UterineZeroCellFactory3d.cpp
@@ -1,4 +1,4 @@
-#include "../include/UterineZeroCellFactory3d.hpp"
+#include "../../include/factories/UterineZeroCellFactory3d.hpp"
 #include "Exception.hpp"
 
 UterineZeroCellFactory3d::UterineZeroCellFactory3d() :
